+++ To-do List - Model View Controller Demo/To-do List - Model View Controller Demo/TodoController.h
@@ -4,25 +4,19 @@
 #include <iostream>
 #include <vector>
 #include <string>
+#include "TodoView.h"
+#include "TodoModel.h"
 
 using namespace std;
 
-class TodoList {
-	vector<pair<string, string>> list;
-public: 
-	void AddtoList(pair<string, string>* p) {
-		list.push_back(*p);
-	}
-};
-
 class TodoController {
-	TodoList* list;
+	TodoView view;
+	TodoModel model;
 public:
-	void addNewItem(string s, string t) {
-		pair<string, string>* p;
-		p->first = s;
-		p->second = t;
-		list->AddtoList(p);
+	void run(){
+		do {
+
+		}
 	}
 }
