+++ To-do List - Model View Controller Demo/To-do List - Model View Controller Demo/TodoView.h
@@ -2,8 +2,30 @@
 #define TODOVIEW_h
 
 #include <iostream>
+#include <vector>
 
 using namespace std;
+
+class TodoList {
+	vector<pair<string, string>> list;
+public:
+	void AddtoList(pair<string, string>* p) {
+		list.push_back(*p);
+	}
+};
+
+class TodoView {
+	TodoList* list;
+public:
+	void addNewItem(string s, string t) {
+		pair<string, string>* p;
+		p->first = s;
+		p->second = t;
+		list->AddtoList(p);
+	}
+
+	void 
+}
